libamba, ipc: Move NodeMetadataFFI to libamba, restructure export of ffi crate
