+++ crates/ipc/src/lib.rs
@@ -2,6 +2,8 @@ mod graph;
 mod ipc;
 mod metadata;
 
-pub use metadata::{NodeMetadata, NodeMetadataFFI};
-
-pub use crate::ipc::*;
+pub use crate::{
+	graph::{GraphIpc, GraphIpcBuilder},
+	ipc::{new_wrapping, IpcError, IpcMessage, IpcRx, IpcTx},
+	metadata::NodeMetadata,
+};
