+++ crates/ipc/src/metadata.rs
@@ -2,33 +2,11 @@ use std::num::NonZeroU64;
 
 use serde::{Deserialize, Serialize};
 
-#[repr(C)]
-pub struct NodeMetadataFFI {
-	pub symbolic_state_id: u32,
-	pub basic_block_vaddr: u64,
-	pub basic_block_generation: u64,
-}
-
 #[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq)]
 pub struct NodeMetadata {
 	pub symbolic_state_id: u32,
 	pub basic_block_vaddr: Option<NonZeroU64>,
 	pub basic_block_generation: Option<NonZeroU64>,
-}
-
-impl From<NodeMetadataFFI> for NodeMetadata {
-	fn from(value: NodeMetadataFFI) -> Self {
-		let NodeMetadataFFI {
-			symbolic_state_id,
-			basic_block_vaddr,
-			basic_block_generation,
-		} = value;
-		NodeMetadata {
-			symbolic_state_id,
-			basic_block_vaddr: basic_block_vaddr.try_into().ok(),
-			basic_block_generation: basic_block_generation.try_into().ok(),
-		}
-	}
 }
 
 impl From<u32> for NodeMetadata {
